fix(parsers): preserve fractional shares and fix Transfer In status
- compute_status: add Transfer In snapshot-reset and Reinvest Shares
  handling, matching every other analysis function — prevents transfer-
  only positions from being permanently stuck as Inconsistent
- robinhood.py, stockpile.py, fidelity.py: parse qty as float instead
  of int(float(...)) so DRIP fractional shares are preserved, matching
  the schwab.py change in the PR
- schwab.py: add math.isfinite guard to reject inf/nan quantities from
  crafted CSV inputs
- report.py, tabs/portfolio.py: format shares with :g to display whole-
  number floats as "100" rather than "100.0" in UI labels and HTML

+++ shared/stocks_shared/parsers/stockpile.py
@@ -40,12 +40,12 @@ def _parse_float(s: str) -> float | None:
         return None
 
 
-def _parse_qty(s: str) -> int | None:
+def _parse_qty(s: str) -> float | None:
     s = s.strip()
     if not s:
         return None
     try:
-        return abs(int(float(s)))
+        return abs(float(s))
     except ValueError:
         return None
